requirements.txt
langchain-core 
langchain 
langchain-openai
langgraph
openai
python-dotenv 
pandas 
tiktoken
